requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.23
pydantic==2.8.2
PyYAML==6.0.1
